renamed entityTable and entityList components
