--- ui/src/components/Entity/EntityListManager.scss
@@ -1,20 +1,0 @@
-@import "app/variables.scss";
-@import "app/mixins.scss";
-
-.EntityListManager {
-  .bp3-button-group {
-    margin-bottom: $aleph-content-padding;
-  }
-
-  .align-middle {
-    vertical-align: middle;
-  }
-
-  .Count {
-    @include rtlSupportInvertedProp(margin, left, 5px, null);
-  }
-
-  .data-grid-container {
-    min-height: 300px;
-  }
-}
