+++ ui/src/components/EntityTable/EntityTableViews.scss
@@ -1,7 +1,7 @@
 @import "app/variables.scss";
 @import "app/mixins.scss";
 
-.EntityListViews {
+.EntityTableViews {
   &__tabs {
     .bp3-tab-list {
       // margin-right: $aleph-content-padding;
